+++ packages/terra-application/src/application-status-overlay/ApplicationStatusOverlayProvider.jsx
@@ -26,10 +26,21 @@ const propTypes = {
    * Prop to not render scroll in  Powerchart.
    */
   noScroll: PropTypes.bool,
+  /**
+   * @private
+   * @type {number}
+   * @default 0
+   * Prop to control the z-index of overlay provider.
+   */
+  zIndex: PropTypes.number,
+};
+
+const defaultProps = {
+  zIndex: 0,
 };
 
 const ApplicationStatusOverlayProvider = ({
-  children, scrollRefCallback, noScroll, ...customProps
+  children, scrollRefCallback, noScroll, zIndex, ...customProps
 }) => {
   const [registeredStatusOverlay, setRegisteredStatusOverlay] = React.useState({});
 
@@ -93,6 +104,7 @@ const ApplicationStatusOverlayProvider = ({
   }, [registeredStatusOverlayKeys]);
 
   let className = cx('container');
+
   if (customProps.className) {
     className = [className, customProps.className].join(' ');
   }
@@ -111,7 +123,8 @@ const ApplicationStatusOverlayProvider = ({
   const scrollClass = cx({ 'remove-scroll': noScroll });
 
   return (
-    <div {...customProps} className={className}>
+    // eslint-disable-next-line react/forbid-dom-props
+    <div {...customProps} className={className} style={{ zIndex }}>
       {statusView}
       <div data-status-overlay-container-content ref={containerRef} className={cx('container-content')}>
         <Scroll className={scrollClass} refCallback={scrollRefCallback}>
@@ -125,5 +138,6 @@ const ApplicationStatusOverlayProvider = ({
 };
 
 ApplicationStatusOverlayProvider.propTypes = propTypes;
+ApplicationStatusOverlayProvider.defaultProps = defaultProps;
 
 export default ApplicationStatusOverlayProvider;